# Cache multi-IR info, particularly include_in_dumps