--- a/toolchain/check/check.cpp
+++ b/toolchain/check/check.cpp
@@ -322,19 +322,18 @@ static auto BuildApiMapAndDiagnosePackaging(
   return api_map;
 }
 
-auto CheckParseTrees(llvm::MutableArrayRef<Unit> units, bool prelude_import,
-                     llvm::IntrusiveRefCntPtr<llvm::vfs::FileSystem> fs,
-                     llvm::raw_ostream* vlog_stream, bool fuzzing) -> void {
+auto CheckParseTrees(
+    llvm::MutableArrayRef<Unit> units,
+    llvm::ArrayRef<Parse::GetTreeAndSubtreesFn> tree_and_subtrees_getters,
+    bool prelude_import, llvm::IntrusiveRefCntPtr<llvm::vfs::FileSystem> fs,
+    llvm::raw_ostream* vlog_stream, bool fuzzing) -> void {
   // UnitAndImports is big due to its SmallVectors, so we default to 0 on the
   // stack.
-  llvm::SmallVector<UnitAndImports, 0> unit_infos;
-  llvm::SmallVector<Parse::GetTreeAndSubtreesFn> tree_and_subtrees_getters;
-  unit_infos.reserve(units.size());
-  tree_and_subtrees_getters.reserve(units.size());
-  for (auto [i, unit] : llvm::enumerate(units)) {
-    unit_infos.emplace_back(SemIR::CheckIRId(i), unit);
-    tree_and_subtrees_getters.push_back(unit.tree_and_subtrees_getter);
-  }
+  llvm::SmallVector<UnitAndImports, 0> unit_infos(
+      llvm::map_range(units, [&](Unit& unit) {
+        return UnitAndImports(
+            &unit, tree_and_subtrees_getters[unit.check_ir_id.index]);
+      }));
 
   Map<ImportKey, UnitAndImports*> api_map =
       BuildApiMapAndDiagnosePackaging(unit_infos);