--- a/toolchain/check/check_unit.h
+++ b/toolchain/check/check_unit.h
@@ -68,18 +68,17 @@ struct UnitAndImports {
     Parse::GetTreeAndSubtreesFn tree_and_subtrees_getter_;
   };
 
-  explicit UnitAndImports(SemIR::CheckIRId check_ir_id, Unit& unit)
-      : check_ir_id(check_ir_id),
-        unit(&unit),
-        err_tracker(*unit.consumer),
-        emitter(&err_tracker, unit.tree_and_subtrees_getter) {}
+  explicit UnitAndImports(Unit* unit,
+                          Parse::GetTreeAndSubtreesFn tree_and_subtrees_getter)
+      : unit(unit),
+        err_tracker(*unit->consumer),
+        emitter(&err_tracker, tree_and_subtrees_getter) {}
 
   auto parse_tree() -> const Parse::Tree& { return unit->sem_ir->parse_tree(); }
   auto source() -> const SourceBuffer& {
     return parse_tree().tokens().source();
   }
 
-  SemIR::CheckIRId check_ir_id;
   Unit* unit;
 
   // Emitter information.
@@ -180,6 +179,7 @@ class CheckUnit {
   auto ProcessNodeIds() -> bool;
 
   UnitAndImports* unit_and_imports_;
+  Parse::GetTreeAndSubtreesFn tree_and_subtrees_getter_;
   // The number of IRs being checked in total.
   int total_ir_count_;
   llvm::IntrusiveRefCntPtr<llvm::vfs::FileSystem> fs_;