--- a/toolchain/language_server/context.cpp
+++ b/toolchain/language_server/context.cpp
@@ -143,16 +143,17 @@ auto Context::File::SetText(Context& context, std::optional<int64_t> version,
     return *tree_and_subtrees_;
   };
   // TODO: Support cross-file checking when multiple files have edits.
-  llvm::SmallVector<Check::Unit> units = {{.consumer = &consumer,
-                                           .value_stores = value_stores_.get(),
-                                           .timings = nullptr,
-                                           .tree_and_subtrees_getter = getter,
-                                           .sem_ir = &sem_ir}};
+  llvm::SmallVector<Check::Unit> units = {{{.consumer = &consumer,
+                                            .value_stores = value_stores_.get(),
+                                            .timings = nullptr,
+                                            .check_ir_id = SemIR::CheckIRId(0),
+                                            .sem_ir = &sem_ir}}};
   llvm::IntrusiveRefCntPtr<llvm::vfs::InMemoryFileSystem> fs =
       new llvm::vfs::InMemoryFileSystem;
   // TODO: Include the prelude.
-  Check::CheckParseTrees(units, /*prelude_import=*/false, fs,
-                         context.vlog_stream(), /*fuzzing=*/false);
+  Check::CheckParseTrees(
+      units, llvm::ArrayRef<Parse::GetTreeAndSubtreesFn>(getter),
+      /*prelude_import=*/false, fs, context.vlog_stream(), /*fuzzing=*/false);
 
   // Note we need to publish diagnostics even when empty.
   // TODO: Consider caching previously published diagnostics and only publishing